Various fixes; it works

+++ include/algorithm.h
@@ -2,6 +2,6 @@
 #define INCLUDE_ALGORITHM_H_
 
 size_t **max_area(const double*, const double*, const int);
-void free_result(int**);
+void free_result(size_t**);
 
 #endif  // INCLUDE_ALGORITHM_H_
